Don't overload operator+.
The typical perception of operator+ is that is doesn't affect the left
argument. If one writes c = b + a it is typically assumed that b is not
affected. But this side effect is wanted by the current implementation
of the overloaded operator+.

To make the side effect more visible, shift the functionality into an
increment method.